355-rename-lagom-dev-mode-components: all related lagom dev mode components are renamed only at the file level. the class name are unchanged.
