+++ dev/service-registry/client-core/src/main/resources/reference.conf
@@ -1,1 +1,1 @@
-akka.discovery.lagom-dev-mode.class= com.lightbend.lagom.internal.registry.DevModeServiceDiscovery
+akka.discovery.lagom-dev-mode.class= com.lightbend.lagom.devmode.internal.registry.DevModeServiceDiscovery
