+++ dev/service-registry/client-core/src/main/scala/com/lightbend/lagom/devmode/internal/registry/LagomDevModeServiceDiscovery.scala
@@ -1,0 +1,47 @@
+/*
+ * Copyright (C) Lightbend Inc. <https://www.lightbend.com>
+ */
+
+package com.lightbend.lagom.devmode.internal.registry
+
+import java.net.{InetAddress, URI}
+
+import akka.actor.ActorSystem
+import akka.discovery.{Discovery, Lookup, ServiceDiscovery}
+import akka.discovery.ServiceDiscovery._
+import com.lightbend.lagom.internal.registry.ServiceRegistryClient
+
+import scala.concurrent.{ExecutionContext, Future, Promise}
+import scala.concurrent.duration.FiniteDuration
+
+private[lagom] class DevModeServiceDiscovery(system: ActorSystem) extends ServiceDiscovery {
+  private val clientPromise = Promise[ServiceRegistryClient]
+
+  private implicit val ec: ExecutionContext = system.dispatcher
+
+  def setServiceRegistryClient(client: ServiceRegistryClient): Unit = clientPromise.success(client)
+
+  override def lookup(lookup: Lookup, resolveTimeout: FiniteDuration): Future[Resolved] =
+    for {
+      client <- clientPromise.future
+      uris   <- client.locateAll(lookup.serviceName, lookup.portName)
+    } yield Resolved(lookup.serviceName, uris.map(toResolvedTarget))
+
+  private def toResolvedTarget(uri: URI) =
+    ResolvedTarget(
+      uri.getHost,
+      optionalPort(uri.getPort),
+      // we don't have the InetAddress, but instead of using None
+      // we default to localhost as such we can use it for Akka Cluster Bootstrap eventually
+      address = Some(InetAddress.getLocalHost)
+    )
+
+  private def optionalPort(port: Int): Option[Int] = if (port < 0) None else Some(port)
+}
+
+private[lagom] object DevModeServiceDiscovery {
+  def apply(system: ActorSystem): DevModeServiceDiscovery =
+    Discovery(system)
+      .loadServiceDiscovery("lagom-dev-mode")
+      .asInstanceOf[DevModeServiceDiscovery]
+}
