+++ dev/service-registry/devmode-scaladsl/src/main/scala/com/lightbend/lagom/internal/scaladsl/registry/ServiceRegistration.scala
@@ -6,6 +6,7 @@ package com.lightbend.lagom.internal.scaladsl.registry
 
 import akka.Done
 import akka.actor.CoordinatedShutdown
+import com.lightbend.lagom.devmode.internal.scaladsl.registry.{ServiceRegistry, ServiceRegistryService}
 import com.lightbend.lagom.internal.registry.serviceDnsRecords
 import com.lightbend.lagom.scaladsl.api.ServiceInfo
 import com.typesafe.config.Config
@@ -14,7 +15,6 @@ import play.api.Logger
 import scala.concurrent.ExecutionContext
 import scala.util.Failure
 import scala.util.Success
-
 import scala.collection._
 
 class ServiceRegistration(
